+++ projects/ids-enterprise-ng/src/lib/popdown/soho-popdown.directive.ts
@@ -26,6 +26,38 @@ export class SohoPopDownDirective implements AfterViewInit, OnDestroy {
     this.popdownOptions.keepOpen = keepOpen;
   }
 
+   /*
+   * If defined, provides a way to place the popdown against an alternate element.
+   */
+  @Input()
+  set trigger(trigger: any) {
+    this.popdownOptions.trigger = trigger;
+  }
+
+   /*
+   * If true, when the popdown is opened, the first available input/button in its content area will be focused.
+   */
+  @Input()
+  set autoFocus(autoFocus: boolean) {
+    this.popdownOptions.autoFocus = autoFocus;
+  }
+
+   /*
+   * If true, popdown will be toggle soon focused on the popdown trigger.
+   */
+  @Input()
+  set toggleOnFocus(toggleOnFocus: boolean) {
+    this.popdownOptions.toggleOnFocus = toggleOnFocus;
+  }
+
+  /**
+   * Hook to work with tabbing in and out of the popdown.
+   */
+  @Input()
+  set firstLastTab(firstLastTab: any) {
+    this.popdownOptions.firstLastTab = firstLastTab;
+  }
+
   constructor(private element: ElementRef) {}
 
   ngAfterViewInit() {