+++ src/app/editor/editor.demo.ts
@@ -1,6 +1,5 @@
 import {
   Component,
-  OnInit,
   ViewChild,
   AfterViewInit
 } from '@angular/core';
@@ -10,7 +9,6 @@ import { DomSanitizer } from '@angular/platform-browser';
 import {
   SohoEditorComponent
 } from 'ids-enterprise-ng';
-import { AlertDemoComponent } from '../alert/alert.demo';
 
 @Component({
   selector: 'app-editor-demo',
@@ -21,7 +19,7 @@ export class EditorDemoComponent implements AfterViewInit {
 
   @ViewChild(SohoEditorComponent) editor: SohoEditorComponent;
 
-  public model;
+  public model: any;
   public showMultipleEditors = false;
   public showModel = false;
   public editorDisabled = false;
@@ -30,9 +28,10 @@ export class EditorDemoComponent implements AfterViewInit {
   constructor(sanitizer: DomSanitizer) {
     // tslint:disable
     // This text is assumed trusted through the editor since we handle security in our back end and database.
-    // (Or you should be careful here)
+    // Or you should be careful here
     this.model = {
-      editorText: sanitizer.bypassSecurityTrustHtml(`<a href="http://en.wikipedia.org/wiki/e-commerce" class="hyperlink">e-commerce action-items</a>, reintermediate, ecologies paradigms wireless share life-hacks create innovative harness. Evolve solutions rich-clientAPIs synergies harness relationships virtual vertical facilitate end-to-end, wireless, evolve synergistic synergies.</p> <p>Cross-platform, evolve, ROI scale cultivate eyeballs addelivery, e-services content cross-platform leverage extensible viral incentivize integrateAJAX-enabled sticky evolve magnetic cultivate leverage; cutting-edge. Innovate, end-to-end podcasting, whiteboard streamline e-business social; compelling, "cross-media exploit infomediaries innovative integrate integrateAJAX-enabled." Killer interactive reinvent, cultivate widgets leverage morph.</p>`)
+      editorText: sanitizer.bypassSecurityTrustHtml(`<p>Embrace <a href="http://en.wikipedia.org/wiki/e-commerce" class="hyperlink">e-commerce action-items</a>, reintermediate, ecologies paradigms wireless share life-hacks create innovative harness. Evolve solutions rich-clientAPIs synergies harness relationships virtual vertical facilitate end-to-end, wireless, evolve synergistic synergies.</p>
+        <p>Cross-platform, evolve, ROI scale cultivate eyeballs addelivery, e-services content cross-platform leverage extensible viral incentivize integrateAJAX-enabled sticky evolve magnetic cultivate leverage; cutting-edge. Innovate, end-to-end podcasting, whiteboard streamline e-business social; compelling, "cross-media exploit infomediaries innovative integrate integrateAJAX-enabled." Killer interactive reinvent, cultivate widgets leverage morph.</p>`)
     };
 
     // tslint:enable
@@ -40,22 +39,7 @@ export class EditorDemoComponent implements AfterViewInit {
 
   ngAfterViewInit() {
 
-    // Customize the buttons on init
-    this.editor.buttons = {
-      editor: [
-        'header1', 'header2',
-        'separator', 'bold', 'underline', 'strikethrough',
-        'separator', 'foreColor',
-        'separator', 'justifyLeft', 'justifyCenter', 'justifyRight',
-        'separator', 'quote', 'orderedlist', 'unorderedlist',
-        'separator', 'anchor',
-        'separator', 'clearFormatting',
-        'separator', 'source'
-      ],
-      source: [
-        'visual'
-      ]
-    };
+    // Customize the settings on init
 
     this.editor.anchor = {
       url: 'http://www.example.com',
@@ -67,7 +51,7 @@ export class EditorDemoComponent implements AfterViewInit {
 
   }
 
-  onUpdated(event) {
+  onUpdated(event: any) {
     alert(event);
   }